Split the admin-api-listings create shard to stay under 400 lines
Codex review on #1621 flagged create.test.ts (484 lines) for exceeding
the ~400-line target the split was meant to honour. Move the group
validation describe into its own groups.test.ts, leaving create.test.ts
focused on plain creates and date/closes handling. Both shards now sit
under 400 lines; all tests carry over unchanged.
